pyproject.toml
[project]
name = "bentoml-university-score"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bentoml>=1.4.30",
    "joblib>=1.5.2",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "starlette>=0.50.0",
]
